Address comments and move main script into main function for protection
